+++ Dockerfile-alpinebuild
@@ -1,0 +1,10 @@
+FROM ec_golanggrpc_alpinebase
+
+ARG GROUP_ID
+ARG USER_ID
+
+RUN addgroup -g ${GROUP_ID} builder \
+  && adduser -h /home/builder -G builder -D -u ${USER_ID} builder \
+  && mkdir -p /home/builder/src
+
+WORKDIR /home/builder
