+++ buildscripts/clean.sh
@@ -1,13 +1,7 @@
-#!/bin/sh -e
+#!/bin/sh
 # 
 # returns repo to pristine state
 #
-NAME=$( cat name )
-
-rm -rf bin/ \
-       htmlcov/ \
-       pkg/ \
-       pkg-build/
-
-( cd src/${NAME}/api && make clean )
-( cd python && make clean )
+set -e
+git clean -fdX
+./buildscripts/env.sh
