+++ buildscripts/compile.sh
@@ -1,13 +1,9 @@
-#!/bin/sh -e
+#!/bin/sh
 # 
 # Compile all the code
 #
+cd src
+set -e
 NAME=$( cat name )
 
-for m in `find src/${NAME} -path src/${NAME}/vendor -prune -o -name main.go -print`
-do
-	d=$( dirname $m )
-	b=$( basename $d )
-	p=$( echo $d | cut -d'/' -f2- )
-	go build -o bin/${NAME}${b} ${p}
-done
+GOBIN=`pwd`/bin go install ./...
