+++ buildscripts/dependencies.sh
@@ -2,20 +2,8 @@
 # 
 # get dependencies
 #
+cd src
 NAME=$( cat name  )
 
 # generate code from proto definition
-( cd src/${NAME}/api && make )
-( cd python && make dependencies )
-
-# make dependencies
-cd src/${NAME}
-if [ ! -d vendor \
-  -o ! -s Gopkg.toml ]
-then
-	rm -rf Gopkg.*
-	dep init 
-fi
-dep ensure
-dep status 
-cd -
+( cd ${NAME}/api && make )
