+++ buildscripts/functest.sh
@@ -2,11 +2,10 @@
 #
 # Runs functional tests as a package
 #
-export EC_GOLANGGRPC_HOST="alpineapi"
+export EC_GOLANGGRPC_HOST="localhost"
 export EC_GOLANGGRPC_PORT="8080"
 
-CLIENT="bin/ec_golanggrpcclient"
-
+CLIENT="bin/client"
 echo "----> Test no arguments"
 ${CLIENT} 2> /tmp/no_arguments
 cat /tmp/no_arguments
