+++ buildscripts/unittest.sh
@@ -1,26 +1,16 @@
-#!/bin/sh -e
+#!/bin/sh
 # 
 # Unit tests the code and record the coverage
 #
-. ./buildscripts/log
+cd src
+set -e
+rm -rf htmlcov
+mkdir htmlcov
+go test ./... -coverprofile /tmp/c.out
+if [ -s /tmp/c.out ]
+then
+	go tool cover -html=/tmp/c.out -o htmlcov/all.html
+	rm -f /tmp/c.out
+fi
 
-NAME=$( cat name )
-WORKDIR=/src
-
-for d in `find src/${NAME} -path src/${NAME}/vendor -prune -o -type d -print`
-do
-	TESTS=$( ls ${d}/*_test.go 2>/dev/null || true )
-	if [ -n "${TESTS}" ]
-	then
-		sub=$( basename ${d} )
-		pkg=$( echo ${d} | cut -d'/' -f2- )
-		mkdir -p ${WORKDIR}/htmlcov
-		log "Test ${pkg} ..."
-		go test -v -bench . -coverprofile=/tmp/c.out ${pkg}
-		if [ -s /tmp/c.out ]
-		then
-			go tool cover -html=/tmp/c.out -o ${WORKDIR}/htmlcov/${sub}.html
-			rm -f /tmp/c.out
-		fi
-	fi
-done
+# go test --race ./...
