+++ ec_golanggrpc/api/Makefile
@@ -1,4 +1,4 @@
-RELDIR := ../../..
+RELDIR := ../..
 NAME := $(shell cat $(RELDIR)/name )
 PYDIR := $(RELDIR)/python
 PYNAMEDIR := $(PYDIR)/$(NAME)
@@ -11,7 +11,7 @@ compile: api.pb.go $(PYTARGET)
 api.pb.go: api.proto
 	protoc -I . --go_out=plugins=grpc:. $<
 
-$(PYTARGET): $(NAME)/api.proto
+$(PYTARGET): api.proto
 	python3 -m grpc_tools.protoc -I . --python_out=$(PYDIR) --grpc_python_out=$(PYDIR) $<
 
 .PHONY: clean
