+++ ec_golanggrpc/cmd/client/main.go
@@ -1,16 +1,23 @@
 package main
 
 import (
+	"flag"
+	"os"
+
 	log "github.com/sirupsen/logrus"
 
-	"ec_golanggrpc/params"
-	"ec_golanggrpc/server"
+	"github.com/eccles/golanggrpc/ec_golanggrpc/client"
+	"github.com/eccles/golanggrpc/ec_golanggrpc/params"
 )
 
 var parameterMap = params.ParameterMap{
 	"Debug": {
 		Default: "false",
 		Desc:    "Output debug statements to log",
+	},
+	"Host": {
+		Default: "",
+		Desc:    "Hostname",
 	},
 	"Port": {
 		Default: "8080",
@@ -32,12 +39,23 @@ func main() {
 		log.SetLevel(log.DebugLevel)
 	}
 
+	host, err := params.EnvString(parms, "Host")
+	if err != nil {
+		log.Panicf("Error %v", err)
+	}
 	port, err := params.EnvInt(parms, "Port")
 	if err != nil {
 		log.Panicf("Error %v", err)
 	}
+	cmd := flag.String("c", "Echo", "Command to execute")
 
+	log.Infof("Cmd %s", *cmd)
 	log.Info("Debug ", debug)
+	log.Info("Host ", host)
 	log.Info("Port ", port)
-	server.Execute(port)
+	log.Info("Execute")
+
+	if !client.Execute(host, port, cmd) {
+		os.Exit(1)
+	}
 }
