+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/eccles/golanggrpc
+
+go 1.12
+
+require (
+	github.com/golang/protobuf v1.3.1
+	github.com/sirupsen/logrus v1.4.0
+	golang.org/x/net v0.0.0-20190328230028-74de082e2cca
+	google.golang.org/grpc v1.19.1
+)
