+++ python/.gitignore
@@ -1,0 +1,6 @@
+api_pb2*.py
+__pycache__
+build/
+dist/
+ec_golanggrpc.egg-info/
+
