--- python/buildscripts/dependencies.sh
@@ -1,5 +1,0 @@
-#!/bin/sh -ex
-# 
-# get requirements
-#
-pip3 install -r requirements-dev.txt
