+++ python/buildscripts/wheel.sh
@@ -1,7 +1,8 @@
-#!/bin/sh -ex
+#!/bin/sh
 # 
 # make wheel package
 #
+set -e
 PYTHON=python3
 SETUP="${PYTHON} setup.py"
 ${SETUP} bdist_wheel
