--- python/requirements-dev.txt
@@ -1,3 +1,0 @@
-#
--r requirements.txt
-wheel
