+++ python/setup.py
@@ -7,11 +7,9 @@ from collections import namedtuple
 import os
 import re
 from setuptools import setup, find_packages
-from pip._internal.req import parse_requirements
 
 from ec_golanggrpc import about
 
-requirements = parse_requirements('requirements.txt', session=False)
 NAME = open('name', 'rt').read().strip()
 DESC = open('README.md').read()
 
@@ -24,5 +22,5 @@ setup(
     description=about.__description__,
     long_description=DESC,
     packages=find_packages(exclude=['build']),
-    install_requires=[str(r.req) for r in requirements],
+    install_requires=['grpcio', 'protobuf'],
 )
