--- src/ec_golanggrpc/api/api.proto
@@ -1,1 +1,0 @@
-ec_golanggrpc/api.proto
